requirements.txt
astroid==2.15.2
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.7
dill==0.3.6
djlint==1.19.16
EditorConfig==0.12.3
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
html-tag-names==0.1.2
html-void-elements==0.1.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsbeautifier==1.14.7
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mysql-connector==2.2.9
mysql-connector-python==8.0.32
mysql-connector-python-rf==2.2.2
pathspec==0.11.1
platformdirs==3.2.0
protobuf==3.20.3
pylint==2.17.2
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.7
PyYAML==6.0
regex==2022.10.31
six==1.16.0
SQLAlchemy==2.0.9
tomli==2.0.1
tomlkit==0.11.7
tqdm==4.65.0
typing_extensions==4.5.0
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
